refactor(client/AdminPanel/components): refactor components
