+++ client/src/App.tsx
@@ -6,29 +6,29 @@ import Signup from './components/Signup/Signup';
 import Layout from './components/Layout/Layout';
 import RequireAuth from './components/RequireAuth.ts/RequireAuth';
 import { Welcome } from './components/Welcome/Welcome';
-import UsersList from './components/UsersList/UsersList';
+import UsersList from './components/AdminPanel/UsersList/UsersList';
 import { useGetUserAuthQuery } from './slices/authentication/authApiSlice';
-import CreateCategory from './components/CreateCategory/CreateCategory';
+import CreateCategory from './components/AdminPanel/CreateCategory/CreateCategory';
 import UpdateEvent from './components/UpdateEvent/UpdateEvent';
 import Unauthorized from './components/Unauthorized/Unauthorized';
 import ROLES_LIST from './slices/authentication/rolesList';
 import CreateEvent from './components/CreateEvent/CreateEvent';
 import Loading from './components/Loading/Loading';
 import Favorites from './components/Favorites/Favorites';
-import Updateuser from './components/UpdateUser/UpdateUser';
-import EventsConfig from './components/EventsConfig/EventsConfig';
+import Updateuser from './components/AdminPanel/UpdateUser/UpdateUser';
+import EventsConfig from './components/AdminPanel/EventsConfig/EventsConfig';
 import CreateOrganization from './components/CreateOrganization/CreateOrganization';
-import OrganizationList from './components/OrganizationList/OrganizationList';
-import UpdateOrganization from './components/UpdateOrganization/updateOrganization';
+import OrganizationList from './components/AdminPanel/OrganizationList/OrganizationList';
+import UpdateOrganization from './components/AdminPanel/UpdateOrganization/updateOrganization';
 import UserPurchaseDetail from './components/UserPurchaseDetail/UserPurchaseDetail';
-import DetailPayment from './components/DetailPayment/DetailPayment';
-import UpdateRol from './components/UpdateRol/UpdateRol';
+import DetailPayment from './components/AdminPanel/DetailPayment/DetailPayment';
+import UpdateRol from './components/AdminPanel/UpdateRol/UpdateRol';
 import OrganizationEvents from './components/OrganizationEvents/OrganizationEvents';
 import EventLocations from './components/EventLocations/EventLocations';
-import DetailProessPayment from './components/DetailPayment/DetailProcessPayment';
+import DetailProessPayment from './components/AdminPanel/DetailPayment/DetailProcessPayment';
 import UserData from './components/UserData/UserData';
 
-import UnbanUser from './components/UnbanUser/UnbanUser';
+import UnbanUser from './components/AdminPanel/UnbanUser/UnbanUser';
 import ChangePassword from './components/ChangePassword/ChangePassword';
 import CategoriesList from './components/AdminPanel/CategoriesList/CategoriesList';
 import UpdateCategory from './components/AdminPanel/UpdateCategories/UpdateCategories';
