+++ client/src/components/AdminPanel/EventsConfig/event-config.module.css
